requirements.txt
azure-functions
azure-storage
requests
pytz